archipelago: Add support for nested locking

+++ xseg/tools/archipelago/archipelago/common.py
@@ -518,6 +518,21 @@ def construct_peers():
 def get_random_vlmc_port():
     return random.randint(config['VTOOL_START'], config['VTOOL_END'])
 
+acquired_locks = {}
+
+def get_lock(lock_file):
+    while True:
+        try:
+            fd = os.open(lock_file, os.O_CREAT | os.O_EXCL | os.O_WRONLY)
+            break
+        except OSError, (err, reason):
+            print >> sys.stderr, lock_file, reason
+            if err == errno.EEXIST:
+                time.sleep(0.2)
+            else:
+                raise OSError(err, lock_file + ' ' + reason)
+    return fd
+
 def exclusive(get_port=False):
     port = None
     if get_port:
@@ -538,21 +553,23 @@ def exclusive(get_port=False):
                 lock_file = os.path.join(LOCK_PATH, VLMC_LOCK_FILE + '_' + str(port))
             else:
                 lock_file = os.path.join(LOCK_PATH, VLMC_LOCK_FILE)
-            while True:
-                try:
-                    fd = os.open(lock_file, os.O_CREAT | os.O_EXCL | os.O_WRONLY)
-                    break
-                except OSError, (err, reason):
-                    print >> sys.stderr, lock_file, reason
-                    if err == errno.EEXIST:
-                        time.sleep(0.2)
-                    else:
-                        raise OSError(err, lock_file + ' ' + reason)
+            try:
+                depth = acquired_locks[lock_file]
+                if depth == 0:
+                    fd = get_lock(lock_file)
+            except KeyError:
+                acquired_locks[lock_file] = 0
+                fd = get_lock(lock_file)
+
+            acquired_locks[lock_file] += 1
             try:
                 r = fn(*args, port=port, **kwargs)
             finally:
-                os.close(fd)
-                os.unlink(lock_file)
+                acquired_locks[lock_file] -= 1
+                depth = acquired_locks[lock_file]
+                if depth == 0:
+                    os.close(fd)
+                    os.unlink(lock_file)
             return r
 
         return lock
